Fix tsan error
